Added a lot more security. And reproduction.
Genetic algorithm is as follows:
Loop through sum of parents traits.
For each iteration, add rand()%2 to the childs trait.
On average, this is the average of the traits. But it allows for
variation.
Security should be tight enough that it is impossible to improve your
situation by cheating. It is possible to make everything crash. But
that's ok.