Added food, eating.
You can eat now. You get all of the diseases in the food, and all of the
nutrition from the food.
Food can also rot, which decreases the nutrition and increases the risk
of it harming you, which is managed as you getting sick with a disease.
Eventually, it fully biodegrades and vanishes.

Also, now you cannot perform other activities while sleeping or
reproducing. Reproducing takes some time, dictated by a const.