Added ability to throw objects.
Also, food can be infected with disease.

Description of Throwing:
An object can theoretically be thrown. It will gain a speed relative to
the throwers strength and experience in throwing, and the objects weight. Then, every frame, it
will move as much as indicated by its speed, it will slow down due to
weight, and speed up or down if it moves down or up respectively in the
tiles height.
An object will stop if its speed falls to or below 0.

A thrown object will be aimed near where the thrower wants it to go, or
farther, depending on the throwers experience. Once the object reaches
wherever it will end up, it will stop.

TO FIX:
There is no unit public function allowing one to throw. Add one.
Also, the learn functions are not implemented yet, so one cannot gain experience.
Also, when one hits a creature with something, a function in unit is
called, but the function is currently empty.

Finally, I may want to add this feature in the future: when the same type of object is
thrown repeatedly, throwing that object becomes a separate skill with a
separate experience measure.