Added unit weight. Excretion. Other small changes.
Units have a weight. If you are fat, moving takes more energy. Having
lots of energy makes on start converting it to fat. Scaled all energy
related values up.
Being strong takes more living energy. Muscles take energy to maintain.
Weight can be converted to energy if need be, at a limited rate. It will
give back a given percentage of the energy needed to gain weight.
Some of these values are genetic. Spare energy to build weight,
and efficiency on return in particular.
Dying now drops all of your carried items, and produces a corpse object.
The corpse object description is incomplete/incorrect. It ought to
partially be a food, and carry a disease for being raw meat, as well as
any diseases the recently deceased individual had. But for now its not.

Units can excrete. They must whenever hunger rises by a certain amount.
There is some (genetic) amount of time between a need to excrete and
mandatory instant excretion. In that time, you can call the public excrete
method to take care of it.

This is the first time I have a well-named shit() method. Its a highly
accurate desctiption.