Example of boolean expressions in communications.
Incidently, this is what I used to test this system. It may be a little
cluttered, and probably has few comments. boolList and boolList2 are
varients of the same thing, just like boolListUnit and boolListTile are
variants of the same thing.

Note the 18 overloads in foo. 18 overloads
are necessary whenever a non-native type is included in the tuple in
boolList*. Currently, none of these are included in boolListUnit or
boolListTile, but they may be later.