Obstacles now block vision. Security fix.
When there is an obstacle in front of you, i.e. a tree, you won't be
able to see past it. So the leopard hiding behind some bushes will be
able to eat you properly. Note: so far this only applies to seeunit().
Make it apply to seehive() as well.

metabool::get() now checks who you are, and that you aren't spying on
your enemies map vision.
curLoopVar can no longer be reinitialized by someone else.