dded class creature, expanded species. Mod: sight
Organizational changes:
Creature is the common base on animal and unit.
Also, the basic data about animals and units are now stored as part of their species. A lot of the constants about units should be migrated into their species.
 Units are species 0.

Actual changes:
unit sight is no longer global. What a unit sees is stored in the unit in a 2d vector of size lineOfSight*2+1 X linOfSight*2+1, and the point where the unit is standing is (lineofSight, lineOfSight). Seeing is handled by creature, so the same applies to animals.

BUGFIX:
in tile::blocksVision there is an if that checked whether good was false and then may have set good to true. It was supposed to check for good being true. I think.

Nothing was tested/debugged.