Can see something without knowing what.
Also, unit constructor reorganization.
And a small bugfix in the unit::reproduce(int) method.
When seeing, at the edge of your vision there is an area where you have a chance of seeing a creature there perfectly, but have a chance of only knowing that there is something there, without knowing what it is. Signified by the metabool value being 3.
Finally, changed the tile getters so that getting an animal on a tile is as difficult as getting a unit, rather than as difficult as getting random tile data, like whether there's a bush there.