Implemented animal reproduction.
This was done by moving the bulk of the reproduce method to the creature class, along with the variables fetusid and sexuallyMature.
The giveBirth method is pure virtual in the creature class.
A pure virtual createFetus method was added to the creature class, which just creates the new animal/unit, puts it in the proper place, and returns a pointer to the creature part.