Allowed animals to eat food that's an object.
Redesigned animalEatingAbilities so that the part where it decides whether it can eat something on the tile is separate from the rest, which will make it easy to implement an omnivore.
Added rotRate to food, so it doesn't all rot uniformly, and quickly.
Added whether a food is vegetarian to food class, so that it's possible to tell whether a given animal can eat it. This one is public information

Minor changes:
removed eatingBehavior from animal class
Changed object getters, they all return pointers with new dynamically allocated memory.
Added variable to species that determines what sort of height it can climb, i.e. when going from one tile to the next what the max height differential is.
Made possFood in object class a pointer to food.
Changed security on tile::walkable slightly, it now accounts for not being able  to see a tile due to obstacles and stuff.
When a unit eats, the object-food it ate is actually deleted.